+++ kimchi/src/circuits/polynomials/chacha.rs
@@ -254,7 +254,7 @@ where
     F: FftField,
 {
     const ARGUMENT_TYPE: ArgumentType = ArgumentType::Gate(GateType::ChaCha0);
-    const CONSTRAINTS: usize = 5;
+    const CONSTRAINTS: u32 = 5;
 
     fn constraints() -> Vec<E<F>> {
         // a += b; d ^= a; d <<<= 16 (=4*4)
@@ -272,7 +272,7 @@ where
     F: FftField,
 {
     const ARGUMENT_TYPE: ArgumentType = ArgumentType::Gate(GateType::ChaCha1);
-    const CONSTRAINTS: usize = 5;
+    const CONSTRAINTS: u32 = 5;
 
     fn constraints() -> Vec<E<F>> {
         // c += d; b ^= c; b <<<= 12 (=3*4)
@@ -290,7 +290,7 @@ where
     F: FftField,
 {
     const ARGUMENT_TYPE: ArgumentType = ArgumentType::Gate(GateType::ChaCha2);
-    const CONSTRAINTS: usize = 5;
+    const CONSTRAINTS: u32 = 5;
 
     fn constraints() -> Vec<E<F>> {
         // a += b; d ^= a; d <<<= 8  (=2*4)
@@ -308,7 +308,7 @@ where
     F: FftField,
 {
     const ARGUMENT_TYPE: ArgumentType = ArgumentType::Gate(GateType::ChaChaFinal);
-    const CONSTRAINTS: usize = 9;
+    const CONSTRAINTS: u32 = 9;
 
     fn constraints() -> Vec<E<F>> {
         // The last line, namely,
@@ -554,7 +554,7 @@ mod tests {
                 h.insert(Column::Witness(i));
             }
             for i in 0..(lookup_info.max_per_row + 1) {
-                h.insert(Column::LookupSorted(i));
+                h.insert(Column::LookupSorted(i as usize));
             }
             h.insert(Column::Z);
             h.insert(Column::LookupAggreg);
